Added a dtype to all Phenom constants

+++ sage/data/waveform/approximants/helper.py
@@ -44,33 +44,37 @@ class PhenomConstants:
     def __init__(self, device="cuda"):
 
         # Natural numbers
-        self.ZERO = torch.tensor(0.0, device=device)
-        self.ONE = torch.tensor(1.0, device=device)
-        self.THREE = torch.tensor(3.0, device=device)
-        self.FIVE = torch.tensor(5.0, device=device)
-        self.SIX = torch.tensor(6.0, device=device)
-        self.FIFTEEN = torch.tensor(15.0, device=device)
-        self.TWENTY_FOUR = torch.tensor(24.0, device=device)
-        self.FORTY_EIGHT = torch.tensor(48.0, device=device)
-        self.ONE_NINTY_TWO = torch.tensor(192.0, device=device)
+        self.ZERO = torch.tensor(0.0, device=device, dtype=torch.float64)
+        self.ONE = torch.tensor(1.0, device=device, dtype=torch.float64)
+        self.THREE = torch.tensor(3.0, device=device, dtype=torch.float64)
+        self.FIVE = torch.tensor(5.0, device=device, dtype=torch.float64)
+        self.SIX = torch.tensor(6.0, device=device, dtype=torch.float64)
+        self.FIFTEEN = torch.tensor(15.0, device=device, dtype=torch.float64)
+        self.TWENTY_FOUR = torch.tensor(24.0, device=device, dtype=torch.float64)
+        self.FORTY_EIGHT = torch.tensor(48.0, device=device, dtype=torch.float64)
+        self.ONE_NINTY_TWO = torch.tensor(192.0, device=device, dtype=torch.float64)
 
         # Powers of two (except 1)
-        self.TWO = torch.tensor(2.0, device=device)
-        self.FOUR = torch.tensor(4.0, device=device)
-        self.EIGHT = torch.tensor(8.0, device=device)
-        self.SIXTEEN = torch.tensor(16.0, device=device)
-        self.THIRTY_TWO = torch.tensor(32.0, device=device)
-        self.SIXTY_FOUR = torch.tensor(64.0, device=device)
-        self.ONE_TWENTY_EIGHT = torch.tensor(128.0, device=device)
-        self.TWO_FIFTY_SIX = torch.tensor(256.0, device=device)
-        self.FIVE_HUNDRED_AND_TWELVE = torch.tensor(512.0, device=device)
-        self.ONE_THOUSAND_AND_TWENTY_FOUR = torch.tensor(1024.0, device=device)
+        self.TWO = torch.tensor(2.0, device=device, dtype=torch.float64)
+        self.FOUR = torch.tensor(4.0, device=device, dtype=torch.float64)
+        self.EIGHT = torch.tensor(8.0, device=device, dtype=torch.float64)
+        self.SIXTEEN = torch.tensor(16.0, device=device, dtype=torch.float64)
+        self.THIRTY_TWO = torch.tensor(32.0, device=device, dtype=torch.float64)
+        self.SIXTY_FOUR = torch.tensor(64.0, device=device, dtype=torch.float64)
+        self.ONE_TWENTY_EIGHT = torch.tensor(128.0, device=device, dtype=torch.float64)
+        self.TWO_FIFTY_SIX = torch.tensor(256.0, device=device, dtype=torch.float64)
+        self.FIVE_HUNDRED_AND_TWELVE = torch.tensor(
+            512.0, device=device, dtype=torch.float64
+        )
+        self.ONE_THOUSAND_AND_TWENTY_FOUR = torch.tensor(
+            1024.0, device=device, dtype=torch.float64
+        )
 
         # Fractions
-        self.HALF = torch.tensor(0.5, device=device)
-        self.ONE_BY_THREE = torch.tensor(1.0 / 3.0, device=device)
-        self.THREE_BY_TWO = torch.tensor(3.0 / 2.0, device=device)
-        self.FIVE_BY_THREE = torch.tensor(5.0 / 3.0, device=device)
+        self.HALF = torch.tensor(0.5, device=device, dtype=torch.float64)
+        self.ONE_BY_THREE = torch.tensor(1.0 / 3.0, device=device, dtype=torch.float64)
+        self.THREE_BY_TWO = torch.tensor(3.0 / 2.0, device=device, dtype=torch.float64)
+        self.FIVE_BY_THREE = torch.tensor(5.0 / 3.0, device=device, dtype=torch.float64)
 
         # Precomputed
         self.SQRT_6 = torch.sqrt(self.SIX)
@@ -85,19 +89,21 @@ class PhenomConstants:
             setattr(
                 self,
                 name,
-                torch.tensor(value, device=device),
+                torch.tensor(value, device=device, dtype=torch.float64),
             )
 
         ## Physical constants for Pv2
-        self.fM_CUT = torch.tensor(0.2, device=device)
+        self.fM_CUT = torch.tensor(0.2, device=device, dtype=torch.float64)
 
         # QNM Data
-        self._QNMData_a = _QNMData_a.to(device=device)
-        self._QNMData_fdamp = _QNMData_fdamp.to(device=device)
-        self._QNMData_fRD = _QNMData_fRD.to(device=device)
+        self._QNMData_a = _QNMData_a.to(device=device, dtype=torch.float64)
+        self._QNMData_fdamp = _QNMData_fdamp.to(device=device, dtype=torch.float64)
+        self._QNMData_fRD = _QNMData_fRD.to(device=device, dtype=torch.float64)
 
         # Target grid
-        self.QNMData_a = torch.linspace(-1, 1, 500_000, device=device)
+        self.QNMData_a = torch.linspace(
+            -1, 1, 500_000, device=device, dtype=torch.float64
+        )
 
         # Interpolate using your torch cubic function
         self.QNMData_fRD = torch_scipylike_cubic_interp(
